requirements.txt
numpy
Pillow
keras
matplotlib
scikit-learn
h5py
pandas
pyyaml >= 5.1b3
pydot